feat: hybrid wasm32/wasm64 kernel with i64 channel ABI (#269)
## Summary

Migrates the kernel from wasm32 to wasm64 (`wasm64-unknown-unknown`) and
widens the channel ABI from i32 to i64 args, enabling both wasm32 and
wasm64 user processes to run under the same kernel. This is the
foundation for running LP64-dependent programs like MariaDB with correct
`sizeof(long) = 8`.

### Key changes:
- **Kernel wasm64**: Rust kernel targets `wasm64-unknown-unknown` via
`-Z build-std`. All address params (`u32`) widened to `usize`. JS
creates memory64 with `address: 'i64'` and BigInt page counts.
- **Channel ABI widened to i64**: Header grows from 40→72 bytes. 6 args
are now `i64` instead of `i32`. Return value is `i64`. Eliminates
`__SYSCALL_LL_E/O` 64-bit value splitting.
- **Host TypeScript updates**: BigInt conversion for all kernel pointer
params (~35 call sites). `DataView.getBigInt64` for channel arg reads.
`detectPtrWidth()` parses wasm binary to detect wasm32 vs wasm64
processes.
- **Per-process ptrWidth**: `ProcessRegistration` carries `ptrWidth: 4 |
8`. Struct readers (iovec, msghdr, stack_t) handle both layouts. Worker
creates correct memory type and `__channel_base` global type.
- **musl wasm64posix arch**: New `musl-overlay/arch/wasm64posix/` with
LP64 types (`_Addr = long long`, `__LONG_MAX = 0x7fffffffffffffffL`).
Separate `sysroot64/` built by `scripts/build-musl.sh --arch
wasm64posix`.
- **SDK wasm64 toolchain**: `wasm64posix-cc`, `wasm64posix-c++`,
`wasm64posix-configure`, etc. Separate compile/link flags for wasm64
target.
- **All wasm32 programs recompiled**: Updated glue writes `int64_t`
args. `__SYSCALL_LL_E/O` become identity macros. No C source changes
needed.
- **MariaDB wasm64 build**: `build-mariadb.sh` updated for wasm64
toolchain with LP64 struct layouts.
- **wasm64 tests**: `host/test/wasm64.test.ts` verifies `detectPtrWidth`
and hello64 LP64 type sizes.

## Test results
- Cargo: 707 pass, 0 fail
- Vitest: 290 pass, 5 fail (erlang timeout — pre-existing BEAM scheduler
hang)
- libc-test: 0 FAIL, 19 XFAIL
- POSIX: 0 FAIL, 1 XFAIL
- Sortix: 4814 PASS, 0 FAIL, 18 XFAIL, 0 XPASS

+++ crates/kernel/src/fork.rs
@@ -392,14 +392,14 @@ pub fn serialize_fork_state(proc: &Process, buf: &mut [u8]) -> Result<usize, Err
     w.write_i32(proc.terminal.session_id)?;
 
     // ── Program break ──
-    w.write_u32(proc.memory.get_brk())?;
+    w.write_u32(proc.memory.get_brk() as u32)?;
 
     // ── mmap mappings (v5) ──
     let mappings = proc.memory.mappings();
     w.write_u32(mappings.len() as u32)?;
     for m in mappings {
-        w.write_u32(m.addr)?;
-        w.write_u32(m.len)?;
+        w.write_u32(m.addr as u32)?;
+        w.write_u32(m.len as u32)?;
         w.write_u32(m.prot)?;
         w.write_u32(m.flags)?;
     }
@@ -710,7 +710,7 @@ pub fn deserialize_fork_state(buf: &[u8], child_pid: u32) -> Result<Process, Err
     // ── Program break ──
     let program_break = r.read_u32()?;
     let mut memory = MemoryManager::new();
-    memory.set_brk(program_break);
+    memory.set_brk(program_break as usize);
 
     // ── mmap mappings (v5) ──
     if r.remaining() >= 4 {
@@ -720,8 +720,8 @@ pub fn deserialize_fork_state(buf: &[u8], child_pid: u32) -> Result<Process, Err
         }
         let mut mappings = Vec::with_capacity(mapping_count);
         for _ in 0..mapping_count {
-            let addr = r.read_u32()?;
-            let len = r.read_u32()?;
+            let addr = r.read_u32()? as usize;
+            let len = r.read_u32()? as usize;
             let prot = r.read_u32()?;
             let flags = r.read_u32()?;
             mappings.push(MappedRegion { addr, len, prot, flags });
@@ -1033,7 +1033,7 @@ pub fn serialize_exec_state(proc: &Process, buf: &mut [u8]) -> Result<usize, Err
     w.write_i32(proc.terminal.session_id)?;
 
     // ── Program break ──
-    w.write_u32(proc.memory.get_brk())?;
+    w.write_u32(proc.memory.get_brk() as u32)?;
 
     // ── Patch total_size ──
     let total = w.pos as u32;
@@ -1224,7 +1224,7 @@ pub fn deserialize_exec_state(buf: &[u8], pid: u32) -> Result<Process, Errno> {
     // ── Program break ──
     let program_break = r.read_u32()?;
     let mut memory = MemoryManager::new();
-    memory.set_brk(program_break);
+    memory.set_brk(program_break as usize);
 
     Ok(Process {
         pid,
